services/sms_service.py
# services/sms_service.py
"""
SMS communication service for customer outreach
Copyright (c) 2024 [Your Name]. All rights reserved.
"""
import requests
import logging
from config import SMS_API_KEY, SMS_API_URL

class SMSService:
    @staticmethod
    def send_sms(phone_number, message, sender_id="SUPERMART"):
        """
        Send SMS to customer phone number
        This is a mock implementation - replace with actual SMS API
        """
        try:
            # Mock SMS sending - replace with actual API call
            logging.info(f"SMS sent to {phone_number}: {message}")
            
            # Example API call structure (uncomment and modify for real implementation)
            # payload = {
            #     'api_key': SMS_API_KEY,
            #     'to': phone_number,
            #     'message': message,
            #     'sender_id': sender_id
            # }
            # response = requests.post(SMS_API_URL, json=payload)
            # return response.status_code == 200
            
            return True  # Mock success
            
        except Exception as e:
            logging.error(f"Error sending SMS: {e}")
            return False

    @staticmethod
    def send_promotional_sms(customer_list, message):
        """Send promotional SMS to multiple customers"""
        success_count = 0
        failed_count = 0
        
        for customer in customer_list:
            if customer.phone:
                if SMSService.send_sms(customer.phone, message):
                    success_count += 1
                else:
                    failed_count += 1
        
        return success_count, failed_count

    @staticmethod
    def send_low_stock_alert(phone_number, product_name, current_stock):
        """Send low stock alert SMS to manager"""
        message = f"LOW STOCK ALERT: {product_name} has only {current_stock} units remaining. Please reorder."
        return SMSService.send_sms(phone_number, message)

    @staticmethod
    def send_transaction_receipt(phone_number, transaction_number, total_amount):
        """Send transaction receipt via SMS"""
        message = f"Receipt: Transaction {transaction_number} completed. Total: ${total_amount:.2f}. Thank you!"
        return SMSService.send_sms(phone_number, message)


# services/call_service.py
"""
Voice call service for customer communication
Copyright (c) 2024 [Your Name]. All rights reserved.
"""
import requests
import logging
from config import CALL_API_KEY, CALL_API_URL

class CallService:
    @staticmethod
    def make_call(phone_number, message_text, voice_type="female"):
        """
        Make automated voice call to customer
        This is a mock implementation - replace with actual Voice API
        """
        try:
            # Mock call implementation
            logging.info(f"Voice call made to {phone_number}: {message_text}")
            
            # Example API call structure (uncomment and modify for real implementation)
            # payload = {
            #     'api_key': CALL_API_KEY,
            #     'to': phone_number,
            #     'message': message_text,
            #     'voice': voice_type
            # }
            # response = requests.post(CALL_API_URL, json=payload)
            # return response.status_code == 200
            
            return True  # Mock success
            
        except Exception as e:
            logging.error(f"Error making call: {e}")
            return False

    @staticmethod
    def call_customer_for_payment_reminder(phone_number, customer_name, amount_due):
        """Call customer for payment reminder"""
        message = f"Hello {customer_name}, this is a friendly reminder about your outstanding payment of ${amount_due:.2f}. Please visit our store or call us back."
        return CallService.make_call(phone_number, message)

    @staticmethod
    def call_customer_for_promotion(phone_number, customer_name, promotion_details):
        """Call customer with promotional offer"""
        message = f"Hello {customer_name}, we have a special offer for you: {promotion_details}. Visit us today!"
        return CallService.make_call(phone_number, message)

    @staticmethod
    def emergency_call_notification(phone_number, emergency_message):
        """Make emergency notification call"""
        return CallService.make_call(phone_number, emergency_message, voice_type="urgent")
