requirements.txt
mysql-connector-python==8.2.0
python-dotenv==1.0.0
Pillow>=10.4.0
